+++ src/lib/util.ts
@@ -151,3 +151,23 @@ export async function worker(params, callback) {
     workerApi.onmessage = e => resolve(e.data)
   })
 }
+
+export function ensureScriptLoaded(scriptPath: string) {
+  return new Promise<void>(resolve => {
+    const scriptBase = scriptPath.split('?')[0]
+    if (document.querySelector(`script[src^="${scriptBase}"]`)) {
+      resolve()
+      return
+    }
+    const script = document.createElement('script')
+    script.src = `${scriptPath}?v=${__APP_VERSION__}`
+    script.onload = () => resolve()
+    document.head.appendChild(script)
+  })
+}
+export function ensureJszipLoaded() {
+  return ensureScriptLoaded('assets/external/jszip.min.js')
+}
+export function ensureMermaidLoaded() {
+  return ensureScriptLoaded('assets/external/mermaid.min.js')
+}